refactor: separate the context for CheckboxGroup and RadioGroup
