+++ packages/react-styled-core/src/CheckboxGroup/context.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const GroupContext = React.createContext({
+  disabled: false,
+  size: undefined,
+  value: undefined,
+  variantColor: undefined,
+  onChange: (e) => {},
+});
+
+export const useGroupContext = () => React.useContext(GroupContext);
+
+export default GroupContext;
